KS_OMEGA_Orchestrator/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
orjson==3.10.7
rich==13.9.2
pandas==2.2.3
streamlit==1.39.0
sqlite-utils==3.36
